Homebridge compatibility
Hi, I had hardtime with my integration and found out on other curtains that the way home bridge (therefore HomeKit) faces status is the "open amount" - that means 80% is 80% open not closed. I changed some of the references here (still using magic number 100, not variable) and it worked. I still didn't have time to see where it reports the status to make sure it's always correct (even during movement). Also, the open/close action is triggering some default command, which works, but sets the numbers to something different. I also saw on other drivers a property to use or not this inversion, which could be a great improvement.